fix: major improvements to audio streaming, material editing, and component rendering
Audio System Fixes:
- Fix critical audio looping bug on first Play Mode entry by increasing buffer count from 6 to 12
- Implement Unity-style streaming pipeline with immediate underrun recovery
- Add comprehensive buffer cleaning at 4 critical points (OnAttached, Play, Stop, StartStreaming)
- Refactor streaming thread to batch buffer processing for better performance
- Ensure OpenAL sources are properly cleaned when reused to prevent residual data

Material Editing System Rewrite:
- Complete rewrite of MaterialAssetInspector for fluid, real-time UX
- Single in-memory material as source of truth to prevent value resets
- Synchronous saves for immediate operations (textures) with debounced auto-save for sliders
- Live preview updates while editing parameters
- Fixed mini-freezes caused by excessive disk I/O and async race conditions

Component System Improvements:
- Fix MeshRendererComponent.HasMeshToRender() to respect Enabled checkbox
- Fix Terrain.Render() to skip rendering when disabled
- Update all render paths in ViewportRenderer to check HasMeshToRender()

Rendering Fixes:
- Fix IBL lighting inversion by flipping X-coordinate in IBL.glsl for skybox HDR
- Correct irradiance and prefiltered environment map sampling orientation

Inspector UI:
- Remove playback progress bar from AudioSourceInspector for cleaner UI
- Add GlassMaterialInspector and TerrainMaterialInspector
- Remove obsolete WaterComponentInspector and WaterMaterialInspector

Engine Architecture:
- Add PlayModeTransitionManager with 7-phase robust transition pipeline
- Add SystemHandlers interface for play mode system lifecycle
- Add MainThreadInvoker utility for thread-safe operations

Asset System:
- Add GlassMaterialAsset for glass rendering materials
- Remove deprecated WaterMaterialAsset
- Remove WaterComponent (consolidated into shader system)

+++ Engine/Rendering/Shaders/Includes/Common.glsl
@@ -33,7 +33,7 @@ layout(std140, binding=0) uniform Global {
     vec3 uFogColor; float uFogStart;
     float uFogEnd; vec3 _pad16;
 
-    int uClipPlaneEnabled; float _pad17; float _pad18; float _pad19;
+    float uClipPlaneEnabled; float _pad17; float _pad18; float _pad19;
     vec4 uClipPlane; // plane equation: normal.xyz, d
 };
